modyfikacje do 6 zadan

+++ src/main/java/pl/wsei/store/service/BasketService.java
@@ -11,14 +11,26 @@ public class BasketService {
 
     private static EntityManagerFactory emf = Persistence.createEntityManagerFactory("pu");
 
-    public void buyItem(String item) {
+    public void buyItem(String item, int quantity) {
         EntityManager em = emf.createEntityManager();
         try {
             em.getTransaction().begin();
 
-            Basket basket = new Basket();
-            basket.setItem(item);
-            em.persist(basket);
+            Basket existingItem = em.createQuery("SELECT b FROM Basket b WHERE b.item = :item", Basket.class)
+                    .setParameter("item", item)
+                    .getResultStream()
+                    .findFirst()
+                    .orElse(null);
+
+            if (existingItem != null) {
+                existingItem.setQuantity(existingItem.getQuantity() + quantity);
+                em.merge(existingItem);
+            } else {
+                Basket basket = new Basket();
+                basket.setItem(item);
+                basket.setQuantity(quantity);
+                em.persist(basket);
+            }
 
             em.getTransaction().commit();
         } catch (Exception e) {
@@ -27,6 +39,10 @@ public class BasketService {
         } finally {
             em.close();
         }
+    }
+
+    public void buyItem(String item) {
+        buyItem(item, 1);
     }
 
     public List<Basket> getAllItems() {
@@ -38,5 +54,53 @@ public class BasketService {
             em.close();
         }
         return items;
+    }
+
+    public void clearBasket() {
+        EntityManager em = emf.createEntityManager();
+        try {
+            em.getTransaction().begin();
+            em.createQuery("DELETE FROM Basket").executeUpdate();
+            em.getTransaction().commit();
+        } catch (Exception e) {
+            em.getTransaction().rollback();
+            throw e;
+        } finally {
+            em.close();
+        }
+    }
+
+    public void sellItem(String item, int quantity) {
+        EntityManager em = emf.createEntityManager();
+        try {
+            em.getTransaction().begin();
+
+            Basket existingItem = em.createQuery("SELECT b FROM Basket b WHERE b.item = :item", Basket.class)
+                    .setParameter("item", item)
+                    .getResultStream()
+                    .findFirst()
+                    .orElse(null);
+
+            if (existingItem != null) {
+                int newQuantity = existingItem.getQuantity() - quantity;
+                if (newQuantity <= 0) {
+                    em.remove(existingItem);
+                } else {
+                    existingItem.setQuantity(newQuantity);
+                    em.merge(existingItem);
+                }
+            }
+
+            em.getTransaction().commit();
+        } catch (Exception e) {
+            em.getTransaction().rollback();
+            throw e;
+        } finally {
+            em.close();
+        }
+    }
+
+    public void sellItem(String item) {
+        sellItem(item, 1);
     }
 }
